+++ src/Archive/ArchivePacker.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <vector>
 
-namespace Archives
+namespace Archive
 {
 	class ArchivePacker
 	{
