+++ src/Archive/ArchiveUnpacker.cpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <stdexcept>
 
-namespace Archives
+namespace Archive
 {
 	ArchiveUnpacker::ArchiveUnpacker(const std::string& filename) :
 		m_ArchiveFilename(filename), m_Count(0), m_ArchiveFileSize(0) { }
